% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_exons.R
\name{classify_exons}
\alias{classify_exons}
\title{Classify exons as single, first, inner or last.}
\usage{
classify_exons(input)
}
\arguments{
\item{input}{The name of the downloaded gtf file from GENCODE website}
}
\value{
A data frame with additional column describing exon positions within a transcript
}
\description{
This function takes a gtf file from GENCODE and returns a data frame in the R Global Environment containing an additional column which states the exons position within a transcript; whether they are a single exon, first, inner or last exons.
}
\examples{
df <- load_gtf("gencode.v27.lncRNAs.gtf")
classify_exons(df)
}
\keyword{first}
\keyword{inner}
\keyword{last}
\keyword{single}
