% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminate_redundant_exons.R
\name{eliminate_redundant_exons}
\alias{eliminate_redundant_exons}
\title{Eliminate redundant exons}
\usage{
eliminate_redundant_exons(input)
}
\arguments{
\item{input}{The name of the data frame containing all intron coordinates}
}
\value{
A data frame with unique exons
}
\description{
This function takes a data frame containing exon coordinates and return a data frame containing unique exons and only one assignment from redundant ones.
}
\examples{
eliminate_redundant_exons(classified_exons_df)
}
\seealso{
\code{\link{eliminate_redundant_introns}}
}
\keyword{duplicated}
\keyword{redundant}
