% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminate_redundant_introns.R
\name{eliminate_redundant_introns}
\alias{eliminate_redundant_introns}
\title{Eliminate redundant introns}
\usage{
eliminate_redundant_introns(input)
}
\arguments{
\item{input}{The name of the dataframe containing all intron coordinates}
}
\value{
A dataframe with unique introns
}
\description{
This function takes a dataframe of intron coordinates and return a dataframe containing non-redundant introns and only one assignment from redundant ones.
}
\seealso{
\code{\link{eliminate_redundant_exons}}
}
\keyword{duplicated}
\keyword{redundant}
