% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_5ss_motif.R
\name{extract_5ss_motif}
\alias{extract_5ss_motif}
\title{Extract the donor splice site motif for MaxEntScan web tool}
\usage{
extract_5ss_motif(input,genome)
}
\arguments{
\item{input}{The name of the data frame containing intron coordinates}

\item{genome}{The full genome sequence of a particular species. The default value is the human genome sequence of the hg38 assembly}
}
\value{
A data frame with motif coordinates and their sequences
}
\description{
This function takes a dataframe of intron coordinates and the full genome sequence of a particular species and return a "bed-file" structured data frame containing the motif coordinates, their corresponding transcripts id and intron number, and the motif sequence which are also saved as a fasta file in the working directory. The fasta file can be uploaded to MaxEntScan::5ss web tool for the retrieval of the scores of acceptor splice sites.
}
\examples{
extract_5ss_motif(introns_df, genome=BSgenome.Hsapiens.UCSC.hg38)
}
\seealso{
\code{\link{extract_3ss_motif}}
}
\keyword{donor}
\keyword{maxentscan}
\keyword{ss}
