% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_minus.R
\name{extract_minus}
\alias{extract_minus}
\title{Extract elements on the minus strand from a gtf file.}
\usage{
extract_minus(input, type)
}
\arguments{
\item{input}{The name of the downloaded gtf file from GENCODE website}

\item{type}{A string that specifies the type of element to be reported}
}
\value{
A dataframe of elements on the minus strand
}
\description{
This function takes a gtf file from GENCODE or the dataframe containing extracted introns. It returns as a dataframe a specified element; i.e gene, transcript, exon or intron; that are on the reverse strand
}
\seealso{
\code{\link{extract_plus}}
}
\keyword{negative}
\keyword{strand}
