% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pc.R
\name{extract_pc}
\alias{extract_pc}
\title{Extract protein-coding genes from GENCODE basic annotation file}
\usage{
extract_pc(input)
}
\arguments{
\item{input}{The name of the downloaded basic annoation gtf file from GENCODE website}
}
\value{
A dataframe of protein-coding genes
}
\description{
This function takes a dataframe of basic annotations provided by GENCODE as input. The function extracts from the dataframe all the protein-coding genes and gives as an output a dataframe in the R Global Environment. The output dataframe includes the gene, transcript and exon annotations only.
}
\keyword{protein-coding}
