% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode_compare.R
\name{gencode_compare}
\alias{gencode_compare}
\title{Compare number of annotations between 2 different releases from GENCODE}
\usage{
gencode_compare(df1, df2, type)
}
\arguments{
\item{df1}{The name of the first dataframe of a particular release from GENCODE}

\item{df2}{The name of the second dataframe of a particular release from GENCODE}

\item{type}{A string specifying the genetic element to be compared}
}
\value{
Delta: The difference between annotated numbers of a particular genetic element.
Percentage: the percentage of delta
}
\description{
This function compares the number of annotated genes, transcripts, exons or introns from 2 dataframes. It is used to comapre the different annotations between different releases from GENCODE
}
\keyword{gencode}
\keyword{releases}
