% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gtf.R
\name{load_gtf}
\alias{load_gtf}
\title{Load a gtf file from GENCODE as a dataframe.}
\usage{
load_gtf(input)
}
\arguments{
\item{input}{The name of the downloaded gtf file from GENCODE website}
}
\value{
A dataframe of the gtf file selected
}
\description{
This function combines import with loading the gtf file as a dataframe in the R Global Environment
}
\seealso{
\code{\link[rtracklayer]{import}}
}
\keyword{gencode}
\keyword{gtf}
\keyword{import}
