% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_genes.R
\name{se_genes}
\alias{se_genes}
\title{Extract single-exon genes}
\usage{
se_genes(input)
}
\arguments{
\item{input}{The name of the downloaded gtf file from GENCODE website}
}
\value{
A dataframe of the gene IDs of single-exon transcripts
}
\description{
This function takes a gtf file from GENCODE and returns a data frame in the R Global Environment containing a column of the gene id of the single-exon genes.
}
\examples{
df <- load_gtf("gencode.v27.lncRNAs.gtf")
se_genes(df)
}
\seealso{
\code{\link{se_transcripts}}
}
\keyword{genes}
\keyword{single-exon}
