% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_transcripts.R
\name{se_transcripts}
\alias{se_transcripts}
\title{Extract single-exon transcripts}
\usage{
se_transcripts(input)
}
\arguments{
\item{input}{The name of the downloaded gtf file from GENCODE website}
}
\value{
A dataframe of the transcript ids of single-exon transcripts
}
\description{
This function takes a gtf file from GENCODE and returns a dataframe in the R Global Environment containing a column of the transcripts id of the single-exon transcripts.
}
\seealso{
\code{\link{se_genes}}
}
\keyword{single-exon}
\keyword{transcript}
