% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliced_trans_length.R
\name{spliced_trans_length}
\alias{spliced_trans_length}
\title{Calculate the length of splices transcripts}
\usage{
spliced_trans_length(input)
}
\arguments{
\item{input}{The name of the downloaded  gtf file from GENCODE website}
}
\value{
A dataframe of the transcript ids and their length when spliced
}
\description{
This function takes a dataframe of GENCODE annotations as input. It calculates the length of the transcripts considering only the length of their corresponding exons.
}
\keyword{mature}
\keyword{transcript}
