% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_exon.R
\name{stat_exon}
\alias{stat_exon}
\title{Create a summary statistics table for exons width data}
\usage{
stat_exon(input)
}
\arguments{
\item{input}{The name of the data frame containing exon annotations from GENCODE and the ordinal position of each exon}
}
\value{
A data frame of exons width summary statistics
}
\description{
This function takes a data frame of GENCODE annotations as input. It calculates the summary statistics of exons length taking into consideration their ordinal position. It returns a data frame containing the summarized exons length data data.
}
\examples{
df <- load_gtf("gencode.v27.lncRNAs.gtf")
stat_exon(df)
}
\seealso{
\code{\link{stat_intron}}
}
\keyword{parameters}
\keyword{statistics}
