% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_intron.R
\name{stat_intron}
\alias{stat_intron}
\title{Create a summary statistics table for introns width data}
\usage{
stat_intron(input)
}
\arguments{
\item{input}{The name of the data frame containing introns coordinates and the splice sites of each intron}
}
\value{
A data frame of introns width summary statistics
}
\description{
This function takes a data frame of introns as input. It calculates the summary statistics of introns length taking into consideration their ordinal position and splice sites. The function returns a data frame containing the summarized data.
}
\examples{
stat_intron(intron_df)
}
\seealso{
\code{\link{stat_exon}}
}
\keyword{parameters}
\keyword{statistics}
