% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress_output.R
\name{suppress_output}
\alias{suppress_output}
\title{Suppress the output message of R commands}
\usage{
suppress_output(input)
}
\arguments{
\item{input}{The name of an R command}
}
\value{
The output from a command without displaying any message
}
\description{
This function suppresses all output messages from an R command
}
\examples{
suppress_output(stat_intron(intron_df))
}
