% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_exons.R
\name{classify_exons}
\alias{classify_exons}
\title{Classify exons as single, first, inner or last.}
\usage{
classify_exons(input)
}
\arguments{
\item{input}{The name of the downloaded gtf file from GENCODE website}
}
\value{
A dataframe with additional column describing exon positions within a trnscript
}
\description{
This function takes a gtf file from GENCODE and returns a dataframe in the R Global Environment containing an additional columnn which states the exons position within a transcript; whether they are a single exon, first, inner or last exons.
}
\keyword{first}
\keyword{inner}
\keyword{last}
\keyword{single}
