% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminate_redundant_exons.R
\name{eliminate_redundant_exons}
\alias{eliminate_redundant_exons}
\title{Eliminate redundant exons}
\usage{
eliminate_redundant_exons(input)
}
\arguments{
\item{input}{The name of the dataframe containing all intron coordinates}
}
\value{
A dataframe with unique exons
}
\description{
This function takes a dataframe containing exon coordinates and return a dataframe containing unique exons and only one assignment from redundant ones.
}
\seealso{
\code{\link{eliminate_redundant_introns}}
}
\keyword{duplicated}
\keyword{redundant}
