% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminate_redundant_introns.R
\name{eliminate_redundant_introns}
\alias{eliminate_redundant_introns}
\title{Eliminate redundant introns}
\usage{
eliminate_redundant_introns(input)
}
\arguments{
\item{input}{The name of the data frame containing all intron coordinates}
}
\value{
A data frame with unique introns
}
\description{
This function takes a data frame of intron coordinates and return a data frame containing non-redundant introns and only one assignment from redundant ones.
}
\examples{
eliminate_redundant_introns(introns_df)
}
\seealso{
\code{\link{eliminate_redundant_exons}}
}
\keyword{duplicated}
\keyword{redundant}
