% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_minus.R
\name{extract_minus}
\alias{extract_minus}
\title{Extract elements on the minus strand from a gtf file.}
\usage{
extract_minus(input, type)
}
\arguments{
\item{input}{The name of the downloaded gtf file from GENCODE website}

\item{type}{A string that specifies the type of element to be reported}
}
\value{
A data frame of elements on the minus strand
}
\description{
This function takes a gtf file from GENCODE or the data frame containing extracted introns. It returns as a data frame a specified element; i.e gene, transcript, exon or intron; that are on the reverse strand
}
\examples{
df <- load_gtf("gencode.v27.lncRNAs.gtf")
extract_minus(df, type = "gene")
extract_minus(df, type = "transcript")
extract_minus(df, type = "exon")
extract_minus(introns_df, type = "intron")
}
\seealso{
\code{\link{extract_plus}}
}
\keyword{negative}
\keyword{strand}
