% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pc.R
\name{extract_pc}
\alias{extract_pc}
\title{Extract protein-coding genes from GENCODE basic annotation file}
\usage{
extract_pc(input)
}
\arguments{
\item{input}{The name of the downloaded basic annotation gtf file from GENCODE website}
}
\value{
A data frame of protein-coding genes
}
\description{
This function takes a data frame of basic annotations provided by GENCODE as input. The function extracts from the data frame all the protein-coding genes and gives as an output a data frame in the R Global Environment. The output data frame includes the gene, transcript and exon annotations only.
}
\examples{
df <- load_gtf("gencode.v27.basic.annotation.gtf")
extract_pc(df)
}
\keyword{protein-coding}
