% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_plus.R
\name{extract_plus}
\alias{extract_plus}
\title{Extract elements on the plus strand from a gtf file.}
\usage{
extract_plus(input, type)
}
\arguments{
\item{input}{The name of the downloaded gtf file from GENCODE website or a dataframe of introns}

\item{type}{A string that specifies the type of element to be reported}
}
\value{
A dataframe of elements on the plus strand
}
\description{
This function takes a gtf file from GENCODE or the dataframe of extracted introns. It returns as a dataframe a specified element; i.e gene, transcript, exon or intron; that are on the forward strand
}
\seealso{
\code{\link{extract_minus}}
}
\keyword{positive}
\keyword{strand}
