% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_freq.R
\name{genes_freq}
\alias{genes_freq}
\title{Extract numbers of genes containing single or multiple transcripts}
\usage{
genes_freq(input)
}
\arguments{
\item{input}{The name of the downloaded  gtf file from GENCODE website}
}
\value{
A data frame of number of genes and their percentage
}
\description{
This function takes a data frame of annotations provided by GENCODE as input. It calculates the number of different transcripts of a gene. The output is a data frame containing the number of genes having 1,2,3... or several isoforms and their percentage.
}
\examples{
df <- load_gtf("gencode.v27.lncRNAs.gtf")
genes_freq(df)
}
\seealso{
\code{\link{trans_freq}}
}
